src/main/java/racingcar/Car.java
package racingcar;

import static camp.nextstep.edu.missionutils.Randoms.pickNumberInRange;

public class Car {
    private String name; // final?
    private int position;

    public Car(String name) {
        this.name = name;
        this.position = 0;
    }

    public void move() {
        int randomNum = getRandomNumber();
        if (randomNum >= 4) {
            this.position++;
        }
    }

    public String getName() {
        return name;
    }

    public int getPosition() {
        return position;
    }

    protected int getRandomNumber() {
        return pickNumberInRange(0, 9);
    }


}
